+++ lua/dapui/elements/stacks.lua
@@ -1,16 +1,16 @@
 local config = require("dapui.config")
 local Canvas = require("dapui.render.canvas")
-local lib = require("dapui.lib")
+local util = require("dapui.util")
 
 return function(client)
   local dapui = { stacks = {} }
-  local send_ready = lib.create_render_loop(function()
+  local send_ready = util.create_render_loop(function()
     dapui.stacks.render()
   end)
 
   local threads = require("dapui.components.threads")(client, send_ready)
 
-  local buf = lib.create_buffer("DAP Stacks", {
+  local buf = util.create_buffer("DAP Stacks", {
     filetype = "dapui_stacks",
   })
