+++ lua/dapui/util.lua
@@ -1,9 +1,46 @@
-local async = require("dapui.async")
 local config = require("dapui.config")
+local async = require("dapui.async")
 
 local M = {}
 
 local api = vim.api
+
+---@return function
+function M.create_render_loop(render)
+  local render_cond = async.control.Condvar.new()
+  local pending = false
+
+  async.run(function()
+    while true do
+      if not pending then
+        render_cond:wait()
+      end
+      pending = false
+      xpcall(render, function(msg)
+        local traceback = debug.traceback(msg, 1)
+        M.notify(("Rendering failed: %s"):format(traceback), vim.log.levels.WARN)
+      end)
+      async.util.sleep(10)
+    end
+  end)
+
+  return function()
+    pending = true
+    render_cond:notify_all()
+  end
+end
+
+function M.create_buffer(name, options)
+  local buf = async.api.nvim_create_buf(true, true)
+  options = vim.tbl_extend("keep", options or {}, {
+    modifiable = false,
+  })
+  async.api.nvim_buf_set_name(buf, name)
+  for opt, value in pairs(options) do
+    async.api.nvim_buf_set_option(buf, opt, value)
+  end
+  return buf
+end
 
 function M.round(num)
   if num < math.floor(num) + 0.5 then
@@ -36,16 +73,6 @@ function M.is_uri(path)
   end
 end
 
----@param cb fun(session: table)
-function M.with_session(cb, fail_cb)
-  local session = require("dap").session()
-  if session then
-    cb(session)
-  elseif fail_cb then
-    fail_cb()
-  end
-end
-
 function M.open_buf(bufnr, line, column)
   for _, win in pairs(api.nvim_tabpage_list_wins(0)) do
     if api.nvim_win_get_buf(win) == bufnr then
@@ -70,52 +97,6 @@ function M.open_buf(bufnr, line, column)
       end
     end
   end
-end
-
-function M.jump_to_frame(frame, session, set_frame)
-  if set_frame then
-    session:_frame_set(frame)
-    return
-  end
-  local line = frame.line
-  local column = frame.column
-  local source = frame.source
-  if not source then
-    return
-  end
-
-  if (source.sourceReference or 0) > 0 then
-    local buf = vim.api.nvim_create_buf(false, true)
-    session:request("source", { sourceReference = source.sourceReference }, function(response, err)
-      if err then
-        return
-      end
-      if not response.body.content then
-        M.notify("No source available for frame", vim.log.levels.WARN)
-        return
-      end
-      vim.api.nvim_buf_set_lines(buf, 0, 0, true, vim.split(response.body.content, "\n"))
-      M.open_buf(buf, line, column)
-      vim.api.nvim_buf_set_option(buf, "bufhidden", "delete")
-      vim.api.nvim_buf_set_option(buf, "modifiable", false)
-    end)
-    return
-  end
-
-  if not source.path then
-    M.notify("No source available for frame", vim.log.levels.WARN)
-  end
-
-  local path = source.path
-
-  if not column or column == 0 then
-    column = 1
-  end
-
-  local bufnr =
-    vim.uri_to_bufnr(M.is_uri(path) and path or vim.uri_from_fname(vim.fn.fnamemodify(path, ":p")))
-  vim.fn.bufload(bufnr)
-  M.open_buf(bufnr, line, column)
 end
 
 function M.get_selection(start, finish)
@@ -150,15 +131,6 @@ function M.pretty_name(path)
     path = vim.uri_to_fname(path)
   end
   return vim.fn.fnamemodify(path, ":t")
-end
-
-function M.pop(tbl, key, default)
-  local val = default
-  if tbl[key] then
-    val = tbl[key]
-    tbl[key] = nil
-  end
-  return val
 end
 
 function M.format_error(error)
